chunk: split "column" and "Row" to another source file
