+++ util/chunk/column.go
@@ -1,0 +1,144 @@
+// Copyright 2018 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package chunk
+
+import (
+	"unsafe"
+
+	"github.com/pingcap/tidb/types"
+	"github.com/pingcap/tidb/types/json"
+)
+
+func (c *column) appendDuration(dur types.Duration) {
+	*(*types.Duration)(unsafe.Pointer(&c.elemBuf[0])) = dur
+	c.finishAppendFixed()
+}
+
+func (c *column) appendMyDecimal(dec *types.MyDecimal) {
+	*(*types.MyDecimal)(unsafe.Pointer(&c.elemBuf[0])) = *dec
+	c.finishAppendFixed()
+}
+
+func (c *column) appendNameValue(name string, val uint64) {
+	var buf [8]byte
+	*(*uint64)(unsafe.Pointer(&buf[0])) = val
+	c.data = append(c.data, buf[:]...)
+	c.data = append(c.data, name...)
+	c.finishAppendVar()
+}
+
+func (c *column) appendJSON(j json.BinaryJSON) {
+	c.data = append(c.data, j.TypeCode)
+	c.data = append(c.data, j.Value...)
+	c.finishAppendVar()
+}
+
+type column struct {
+	length     int
+	nullCount  int
+	nullBitmap []byte
+	offsets    []int32
+	data       []byte
+	elemBuf    []byte
+}
+
+func (c *column) isFixed() bool {
+	return c.elemBuf != nil
+}
+
+func (c *column) reset() {
+	c.length = 0
+	c.nullCount = 0
+	c.nullBitmap = c.nullBitmap[:0]
+	if len(c.offsets) > 0 {
+		// The first offset is always 0, it makes slicing the data easier, we need to keep it.
+		c.offsets = c.offsets[:1]
+	}
+	c.data = c.data[:0]
+}
+
+func (c *column) isNull(rowIdx int) bool {
+	nullByte := c.nullBitmap[rowIdx/8]
+	return nullByte&(1<<(uint(rowIdx)&7)) == 0
+}
+
+func (c *column) appendNullBitmap(on bool) {
+	idx := c.length >> 3
+	if idx >= len(c.nullBitmap) {
+		c.nullBitmap = append(c.nullBitmap, 0)
+	}
+	if on {
+		pos := uint(c.length) & 7
+		c.nullBitmap[idx] |= byte(1 << pos)
+	} else {
+		c.nullCount++
+	}
+}
+
+func (c *column) appendNull() {
+	c.appendNullBitmap(false)
+	if c.isFixed() {
+		c.data = append(c.data, c.elemBuf...)
+	} else {
+		c.offsets = append(c.offsets, c.offsets[c.length])
+	}
+	c.length++
+}
+
+func (c *column) finishAppendFixed() {
+	c.data = append(c.data, c.elemBuf...)
+	c.appendNullBitmap(true)
+	c.length++
+}
+
+func (c *column) appendInt64(i int64) {
+	*(*int64)(unsafe.Pointer(&c.elemBuf[0])) = i
+	c.finishAppendFixed()
+}
+
+func (c *column) appendUint64(u uint64) {
+	*(*uint64)(unsafe.Pointer(&c.elemBuf[0])) = u
+	c.finishAppendFixed()
+}
+
+func (c *column) appendFloat32(f float32) {
+	*(*float32)(unsafe.Pointer(&c.elemBuf[0])) = f
+	c.finishAppendFixed()
+}
+
+func (c *column) appendFloat64(f float64) {
+	*(*float64)(unsafe.Pointer(&c.elemBuf[0])) = f
+	c.finishAppendFixed()
+}
+
+func (c *column) finishAppendVar() {
+	c.appendNullBitmap(true)
+	c.offsets = append(c.offsets, int32(len(c.data)))
+	c.length++
+}
+
+func (c *column) appendString(str string) {
+	c.data = append(c.data, str...)
+	c.finishAppendVar()
+}
+
+func (c *column) appendBytes(b []byte) {
+	c.data = append(c.data, b...)
+	c.finishAppendVar()
+}
+
+func (c *column) appendTime(t types.Time) {
+	writeTime(c.elemBuf, t)
+	c.finishAppendFixed()
+}
